+++ nautilus_trader/adapters/betfair/providers.py
@@ -34,6 +34,7 @@ from nautilus_trader.adapters.betfair.constants import BETFAIR_PRICE_PRECISION
 from nautilus_trader.adapters.betfair.constants import BETFAIR_QUANTITY_PRECISION
 from nautilus_trader.adapters.betfair.constants import BETFAIR_VENUE
 from nautilus_trader.adapters.betfair.parsing.common import chunk
+from nautilus_trader.common.config import PositiveFloat
 from nautilus_trader.common.providers import InstrumentProvider
 from nautilus_trader.config import InstrumentProviderConfig
 from nautilus_trader.core.correctness import PyCondition
@@ -52,6 +53,8 @@ class BetfairInstrumentProviderConfig(InstrumentProviderConfig, frozen=True, kw_
     ----------
     account_currency : str
         The Betfair account currency.
+    default_min_notional : PositiveFloat, optional
+        The default minimum notional value for instrument definitions (in account currency).
     event_type_ids : list[int], optional
         The event type IDs to filter for.
     event_ids : list[int], optional
@@ -76,6 +79,7 @@ class BetfairInstrumentProviderConfig(InstrumentProviderConfig, frozen=True, kw_
     """
 
     account_currency: str
+    default_min_notional: PositiveFloat | None = None
     event_type_ids: list[int] | None = None
     event_ids: list[int] | None = None
     market_ids: list[str] | None = None
@@ -150,11 +154,24 @@ class BetfairInstrumentProvider(InstrumentProvider):
             or self._config.max_market_start_time,
         )
 
+        account_currency = Currency.from_str(self._config.account_currency)
+        default_min_notional = (
+            Money(self._config.default_min_notional, account_currency)
+            if self._config.default_min_notional
+            else None
+        )
+
         self._log.info("Creating instruments...")
         instruments = [
             instrument
             for metadata in market_metadata
-            for instrument in make_instruments(metadata, currency=currency, ts_event=0, ts_init=0)
+            for instrument in make_instruments(
+                metadata,
+                currency=currency,
+                ts_event=0,
+                ts_init=0,
+                min_notional=default_min_notional,
+            )
         ]
         for instrument in instruments:
             self.add(instrument=instrument)
@@ -173,6 +190,7 @@ def market_catalog_to_instruments(
     currency: str,
     ts_event: int,
     ts_init: int,
+    min_notional: Money | None,
 ) -> list[BettingInstrument]:
     instruments: list[BettingInstrument] = []
     for runner in market_catalog.runners:
@@ -198,6 +216,7 @@ def market_catalog_to_instruments(
             tick_scheme_name=BETFAIR_TICK_SCHEME.name,
             price_precision=BETFAIR_PRICE_PRECISION,
             size_precision=BETFAIR_QUANTITY_PRECISION,
+            min_notional=min_notional,
             ts_event=ts_event,
             ts_init=ts_init,
             info=msgspec.json.decode(bf_encode(market_catalog).decode()),
@@ -211,6 +230,7 @@ def market_definition_to_instruments(
     currency: str,
     ts_event: int,
     ts_init: int,
+    min_notional: Money | None,
 ) -> list[BettingInstrument]:
     instruments: list[BettingInstrument] = []
     for runner in market_definition.runners:
@@ -240,7 +260,7 @@ def market_definition_to_instruments(
             currency=currency,
             price_precision=BETFAIR_PRICE_PRECISION,
             size_precision=BETFAIR_QUANTITY_PRECISION,
-            min_notional=Money(1, Currency.from_str(currency)),
+            min_notional=min_notional,
             ts_event=ts_event,
             ts_init=ts_init,
             info=msgspec.json.decode(msgspec.json.encode(market_definition)),
@@ -254,6 +274,7 @@ def make_instruments(
     currency: str,
     ts_event: int,
     ts_init: int,
+    min_notional: Money | None = None,
 ) -> list[BettingInstrument]:
     if isinstance(market, MarketCatalogue):
         return market_catalog_to_instruments(
@@ -261,6 +282,7 @@ def make_instruments(
             currency=currency,
             ts_event=ts_event,
             ts_init=ts_init,
+            min_notional=min_notional,
         )
     elif isinstance(market, MarketDefinition):
         return market_definition_to_instruments(
@@ -268,6 +290,7 @@ def make_instruments(
             currency=currency,
             ts_event=ts_event,
             ts_init=ts_init,
+            min_notional=min_notional,
         )
     else:
         # Unreachable unless code changes